change filename and rearrange code
